# docs/openapi/management-v1.openapi.json
{
  "components": {
    "schemas": {
      "CapabilitiesResponse": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "schema_version": {
            "const": "1.4",
            "default": "1.4",
            "title": "Schema Version",
            "type": "string"
          },
          "server_version": {
            "title": "Server Version",
            "type": "string"
          },
          "storage_backend": {
            "title": "Storage Backend",
            "type": "string"
          }
        },
        "required": [
          "server_version",
          "revision",
          "generated_at",
          "storage_backend",
          "capabilities"
        ],
        "title": "CapabilitiesResponse",
        "type": "object"
      },
      "CooldownSyncActionResult": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "cooldown_sync",
            "default": "cooldown_sync",
            "title": "Kind",
            "type": "string"
          },
          "model_cooldowns": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Model Cooldowns",
            "type": "object"
          }
        },
        "required": [
          "model_cooldowns"
        ],
        "title": "CooldownSyncActionResult",
        "type": "object"
      },
...
      "CredentialBatchActionItem": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "enable",
              "disable",
              "permanent_disable",
              "delete",
              "set_remark",
              "enable_preview",
              "disable_preview",
              "enable_credit",
              "disable_credit",
              "quota",
              "errors",
              "test",
              "risk_check",
              "sync_cooldown"
            ],
            "title": "Action",
            "type": "string"
          },
          "filename": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "mode": {
            "enum": [
              "geminicli",
              "antigravity"
            ],
            "title": "Mode",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "action",
          "idempotency_key",
          "mode",
          "filename"
        ],
        "title": "CredentialBatchActionItem",
        "type": "object"
      },
      "CredentialBatchActionRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CredentialBatchActionItem"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "idempotency_key",
          "items"
        ],
        "title": "CredentialBatchActionRequest",
        "type": "object"
      },
      "CredentialBatchActionResponse": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CredentialBatchActionResult"
            },
            "title": "Results",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "schema_version": {
            "const": "1.4",
            "default": "1.4",
            "title": "Schema Version",
            "type": "string"
          },
          "server_version": {
            "title": "Server Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "succeeded",
              "partially_succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "server_version",
          "revision",
          "generated_at",
          "status",
          "results"
        ],
        "title": "CredentialBatchActionResponse",
        "type": "object"
      },
      "CredentialBatchActionResult": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "enable",
              "disable",
              "permanent_disable",
              "delete",
              "set_remark",
              "enable_preview",
              "disable_preview",
              "enable_credit",
              "disable_credit",
              "quota",
              "errors",
              "test",
              "risk_check",
              "sync_cooldown"
            ],
            "title": "Action",
            "type": "string"
          },
          "credential_status": {
            "anyOf": [
              {
                "enum": [
                  "enabled",
                  "disabled",
                  "permanent_disabled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Status"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "mode": {
            "enum": [
              "geminicli",
              "antigravity"
            ],
            "title": "Mode",
            "type": "string"
          },
          "no_change": {
            "title": "No Change",
            "type": "boolean"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviewActionResult"
              },
              {
                "$ref": "#/components/schemas/CreditActionResult"
              },
              {
                "$ref": "#/components/schemas/QuotaActionResult"
              },
              {
                "$ref": "#/components/schemas/ErrorsActionResult"
              },
              {
                "$ref": "#/components/schemas/TestActionResult"
              },
              {
                "$ref": "#/components/schemas/RiskActionResult"
              },
              {
                "$ref": "#/components/schemas/CooldownSyncActionResult"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "side_effects": {
            "items": {
              "$ref": "#/components/schemas/SideEffect"
            },
            "title": "Side Effects",
            "type": "array"
          },
          "status": {
            "enum": [
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "filename",
          "action",
          "status",
          "no_change",
          "credential_status",
          "error"
        ],
        "title": "CredentialBatchActionResult",
        "type": "object"
      },
...
      "CreditActionResult": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "kind": {
            "const": "credit",
            "default": "credit",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "CreditActionResult",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "details"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEntryResult": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 64,
            "title": "Code",
            "type": "string"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorEntryResult",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ErrorsActionResult": {
        "additionalProperties": false,
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ErrorEntryResult"
            },
            "maxItems": 128,
            "title": "Entries",
            "type": "array"
          },
          "kind": {
            "const": "errors",
            "default": "errors",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "entries"
        ],
        "title": "ErrorsActionResult",
        "type": "object"
      },
...
      "PreviewActionResult": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "kind": {
            "const": "preview",
            "default": "preview",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "PreviewActionResult",
        "type": "object"
      },
      "QuotaActionResult": {
        "additionalProperties": false,
        "properties": {
          "captured_at": {
            "title": "Captured At",
            "type": "string"
          },
          "kind": {
            "const": "quota",
            "default": "quota",
            "title": "Kind",
            "type": "string"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/QuotaModelResult"
            },
            "maxItems": 128,
            "title": "Models",
            "type": "array"
          }
        },
        "required": [
          "captured_at",
          "models"
        ],
        "title": "QuotaActionResult",
        "type": "object"
      },
      "QuotaModelResult": {
        "additionalProperties": false,
        "properties": {
          "model_name": {
            "maxLength": 200,
            "title": "Model Name",
            "type": "string"
          },
          "remaining_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Percent"
          },
          "resets_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resets At"
          }
        },
        "required": [
          "model_name",
          "remaining_percent",
          "resets_at"
        ],
        "title": "QuotaModelResult",
        "type": "object"
      },
      "RiskActionResult": {
        "additionalProperties": false,
        "properties": {
          "codes": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Codes",
            "type": "array"
          },
          "kind": {
            "const": "risk",
            "default": "risk",
            "title": "Kind",
            "type": "string"
          },
          "level": {
            "enum": [
              "low",
              "medium",
              "high",
              "unknown"
            ],
            "title": "Level",
            "type": "string"
          }
        },
        "required": [
          "level",
          "codes"
        ],
        "title": "RiskActionResult",
        "type": "object"
      },
      "SideEffect": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "kind": {
            "enum": [
              "credential_state_updated",
              "token_refreshed",
              "google_api_called",
              "cooldown_updated",
              "quota_cache_hit"
            ],
            "title": "Kind",
            "type": "string"
          },
          "occurred": {
            "title": "Occurred",
            "type": "boolean"
          }
        },
        "required": [
          "kind",
          "occurred"
        ],
        "title": "SideEffect",
        "type": "object"
      },
...
      "TestActionResult": {
        "additionalProperties": false,
        "properties": {
          "call_succeeded": {
            "title": "Call Succeeded",
            "type": "boolean"
          },
          "classification": {
            "enum": [
              "success",
              "rate_limited",
              "authentication_failed",
              "rejected",
              "upstream_error",
              "unknown"
            ],
            "title": "Classification",
            "type": "string"
          },
          "kind": {
            "const": "test",
            "default": "test",
            "title": "Kind",
            "type": "string"
          },
          "latency_ms": {
            "anyOf": [
              {
                "maximum": 3600000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "model_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "outcome": {
            "enum": [
              "passed",
              "failed"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "upstream_status": {
            "anyOf": [
              {
                "maximum": 599.0,
                "minimum": 100.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Status"
          }
        },
        "required": [
          "outcome",
          "classification",
          "call_succeeded"
        ],
        "title": "TestActionResult",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "gcli2api Management API",
    "version": "management-schema-1.4"
  },
  "openapi": "3.1.0",
  "paths": {
    "/management/v1/capabilities": {
      "get": {
        "operationId": "capabilities_management_v1_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Capabilities",
        "tags": [
          "Management API v1"
        ]
      }
    },
...
    "/management/v1/credentials/batch-actions": {
      "post": {
        "operationId": "credential_batch_action_management_v1_credentials_batch_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialBatchActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialBatchActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Credential Batch Action",
        "tags": [
          "Management API v1"
        ]
      }
    },
...
  }
}